pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mcp[cli]>=1.6.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]
